Move meta svc test helpers from common test lib into meta svc project
