+++ tracdap-services/tracdap-svc-meta/build.gradle
@@ -78,6 +78,9 @@ dependencies {
 
     // Always pull in the H2 JDBC driver as a test dependency, since it used for unit tests
     testImplementation group: 'com.h2database', name: 'h2', version: "$h2_version"
+
+    // Use Flyway to deploy schemas in test cases
+    testImplementation group: 'org.flywaydb', name: 'flyway-core', version: "$flyway_version"
 }
 
 // Add any plugin dependencies enabled at build time
