+++ tracdap-services/tracdap-svc-meta/src/test/java/org/finos/tracdap/svc/meta/dal/MetadataDalEncodingTest.java
@@ -26,9 +26,9 @@ import java.time.*;
 import java.util.Collections;
 import java.util.UUID;
 
-import org.finos.tracdap.test.meta.IDalTestable;
-import org.finos.tracdap.test.meta.JdbcUnit;
-import org.finos.tracdap.test.meta.JdbcIntegration;
+import org.finos.tracdap.svc.meta.test.IDalTestable;
+import org.finos.tracdap.svc.meta.test.JdbcUnit;
+import org.finos.tracdap.svc.meta.test.JdbcIntegration;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Disabled;
