+++ tracdap-services/tracdap-svc-meta/src/test/java/org/finos/tracdap/svc/meta/dal/MetadataDalSearchTest.java
@@ -22,6 +22,10 @@ import org.finos.tracdap.common.metadata.TypeSystem;
 import org.finos.tracdap.common.metadata.MetadataCodec;
 import org.finos.tracdap.test.meta.*;
 
+import org.finos.tracdap.svc.meta.test.IDalTestable;
+import org.finos.tracdap.svc.meta.test.JdbcIntegration;
+import org.finos.tracdap.svc.meta.test.JdbcUnit;
+
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
